Fixing bad url
```
hint: Using 'master' as the name for the initial branch. This default branch name
hint: is subject to change.
```

No it's not, just because you americans have national issues on how to treat eachother, doesn't mean the rest of the world are having your issues...

We went to FREE TAX paid public schools with no firearms or armed guards...

Signed by @spirillen